iio: humidity: si7020: add support for enabling the heater
All the devices currently supported by this driver have an integrated
on-chip heater that can be configured. The heater can be switched on/off
by toggling a bit in the User Register. This commit adds support for it
by exposing a `heater_enable` file in the sysfs interface. Writing 0 to
the file switches the heater off, writing 1 switches it on.

Signed-off-by: Antoni Pokusiński <apokusinski@o2.pl>